# kenyan-real-estate-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Amenities:
    additionalProperties: true
    type: object
  models.County:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/models.Property'
        type: array
      sub_counties:
        description: Relationships
        items:
          $ref: '#/definitions/models.SubCounty'
        type: array
    type: object
...
  models.Property:
    properties:
      amenities:
        $ref: '#/definitions/models.Amenities'
      availability_date:
        type: string
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      county:
        allOf:
        - $ref: '#/definitions/models.County'
        description: Relationships
      county_id:
        type: integer
      created_at:
        type: string
      deposit_amount:
        type: number
      description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/models.PropertyImage'
        type: array
      is_available:
        type: boolean
      is_furnished:
        type: boolean
      landlord:
        $ref: '#/definitions/models.User'
      landlord_id:
        type: string
      latitude:
        type: number
      location_details:
        type: string
      longitude:
        type: number
      parking_spaces:
        type: integer
      property_type:
        $ref: '#/definitions/models.PropertyType'
      rent_amount:
        type: number
      square_meters:
        type: number
      sub_county:
        $ref: '#/definitions/models.SubCounty'
      sub_county_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      utilities_included:
        $ref: '#/definitions/models.UtilitiesIncluded'
    type: object
  models.PropertyImage:
    properties:
      bytes:
        type: integer
      caption:
        type: string
      created_at:
        type: string
      display_order:
        type: integer
      format:
        type: string
      height:
        type: integer
      id:
        type: string
      image_url:
        type: string
      is_primary:
        type: boolean
      property:
        allOf:
        - $ref: '#/definitions/models.Property'
        description: Relationships
      property_id:
        type: string
      public_id:
        type: string
      secure_url:
        type: string
      width:
        type: integer
    type: object
  models.PropertyType:
    enum:
    - apartment
    - house
    - bedsitter
    - studio
    - maisonette
    - bungalow
    - villa
    - commercial
    type: string
    x-enum-varnames:
    - PropertyTypeApartment
    - PropertyTypeHouse
    - PropertyTypeBedsitter
    - PropertyTypeStudio
    - PropertyTypeMaisonette
    - PropertyTypeBungalow
    - PropertyTypeVilla
    - PropertyTypeCommercial
  models.SubCounty:
    properties:
      county:
        allOf:
        - $ref: '#/definitions/models.County'
        description: Relationships
      county_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/models.Property'
        type: array
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      profile_image_url:
        type: string
      properties:
        description: Relationships
        items:
          $ref: '#/definitions/models.Property'
        type: array
      updated_at:
        type: string
      user_type:
        $ref: '#/definitions/models.UserType'
    type: object
...
  models.UserType:
    enum:
    - landlord
    - tenant
    - agent
    type: string
    x-enum-varnames:
    - UserTypeLandlord
    - UserTypeTenant
    - UserTypeAgent
  models.UtilitiesIncluded:
    additionalProperties: true
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for managing real estate properties in Kenya
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kenyan Real Estate API
  version: "1.0"
paths:
...
  /counties:
    get:
      consumes:
      - application/json
      description: Get a list of all counties in Kenya
      produces:
      - application/json
      responses:
        "200":
          description: List of counties
          schema:
            properties:
              counties:
                items:
                  $ref: '#/definitions/models.County'
                type: array
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all counties
      tags:
      - Location
  /counties/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific county
      parameters:
      - description: County ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: County details
          schema:
            properties:
              county:
                $ref: '#/definitions/models.County'
            type: object
        "400":
          description: Invalid county ID
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: County not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get county by ID
      tags:
      - Location
  /counties/{id}/sub-counties:
    get:
      consumes:
      - application/json
      description: Get all sub-counties within a specific county
      parameters:
      - description: County ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of sub-counties
          schema:
            properties:
              county_id:
                type: integer
              sub_counties:
                items:
                  $ref: '#/definitions/models.SubCounty'
                type: array
            type: object
        "400":
          description: Invalid county ID
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: County not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get sub-counties by county ID
      tags:
      - Location
...
  /sub-counties/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific sub-county
      parameters:
      - description: Sub-county ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sub-county details
          schema:
            properties:
              sub_county:
                $ref: '#/definitions/models.SubCounty'
            type: object
        "400":
          description: Invalid sub-county ID
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Sub-county not found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal server error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get sub-county by ID
      tags:
      - Location
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"